+++ cmd/frontend/graphqlbackend/schema.go
@@ -475,6 +475,11 @@ type CampaignPlan implements Node {
     # The progress status of generating changesets.
     status: BackgroundProcessStatus!
 
+    # DEPRECATED: This field will be removed in 3.15. Please use changesetPlans instead.
+    # The proposed patches ("plans") for the changesets that will be created by the campaign.
+    changesets(first: Int): ChangesetPlanConnection!
+
+    # DEPRECATED: This field will be removed in 3.15. Please use changesetPlans instead.
     # The proposed patches ("plans") for the changesets that will be created by the campaign.
     changesetPlans(first: Int): ChangesetPlanConnection!