+++ enterprise/internal/a8n/resolvers/campaign_plans.go
@@ -58,6 +58,13 @@ func (r *campaignPlanResolver) Status(ctx context.Context) (graphqlbackend.Backg
 	return r.store.GetCampaignPlanStatus(ctx, r.campaignPlan.ID)
 }
 
+// DEPRECATED: Remove in 3.15 in favor of ChangesetPlans.
+func (r *campaignPlanResolver) Changesets(
+	ctx context.Context,
+	args *graphqlutil.ConnectionArgs,
+) graphqlbackend.ChangesetPlansConnectionResolver {
+	return r.ChangesetPlans(ctx, args)
+}
 func (r *campaignPlanResolver) ChangesetPlans(
 	ctx context.Context,
 	args *graphqlutil.ConnectionArgs,